Cleanup frontend face handling
